+++ animation_nodes/utils/handlers.py
@@ -15,8 +15,10 @@ def validCallback(function):
 alwaysHandlers = []
 fileSavePreHandlers = []
 fileLoadPostHandlers = []
+versionUpdateHandlers = []
 addonLoadPostHandlers = []
 frameChangePostHandlers = []
+depsgraphUpdatePostHandlers = []
 
 renderPreHandlers = []
 renderInitHandlers = []
@@ -28,8 +30,10 @@ def eventHandler(event):
         if event == "ALWAYS": alwaysHandlers.append(function)
         if event == "FILE_SAVE_PRE": fileSavePreHandlers.append(function)
         if event == "FILE_LOAD_POST": fileLoadPostHandlers.append(function)
+        if event == "VERSION_UPDATE": versionUpdateHandlers.append(function)
         if event == "ADDON_LOAD_POST": addonLoadPostHandlers.append(function)
         if event == "FRAME_CHANGE_POST": frameChangePostHandlers.append(function)
+        if event == "DEPSGRAPH_UPDATE_POST": depsgraphUpdatePostHandlers.append(function)
 
         if event == "RENDER_INIT": renderInitHandlers.append(function)
         if event == "RENDER_PRE": renderPreHandlers.append(function)
@@ -61,6 +65,11 @@ def loadPost(scene):
         handler()
 
 @persistent
+def versionUpdate(self):
+    for handler in versionUpdateHandlers:
+        handler()
+
+@persistent
 def renderPre(scene):
     for handler in renderPreHandlers:
         handler()
@@ -68,6 +77,11 @@ def renderPre(scene):
 @persistent
 def frameChangedPost(scene, depsgraph):
     for handler in frameChangePostHandlers:
+        handler(scene, depsgraph)
+
+@persistent
+def depsgraphUpdatePost(scene, depsgraph):
+    for handler in depsgraphUpdatePostHandlers:
         handler(scene, depsgraph)
 
 @persistent
@@ -87,8 +101,10 @@ def renderCompleted(scene):
 
 def register():
     bpy.app.handlers.frame_change_post.append(frameChangedPost)
+    bpy.app.handlers.depsgraph_update_post.append(depsgraphUpdatePost)
     bpy.app.timers.register(always, persistent = True)
     bpy.app.handlers.load_post.append(loadPost)
+    bpy.app.handlers.version_update.append(versionUpdate)
     bpy.app.handlers.save_pre.append(savePre)
 
     bpy.app.handlers.render_complete.append(renderCompleted)
@@ -96,12 +112,15 @@ def register():
     bpy.app.handlers.render_cancel.append(renderCancelled)
     bpy.app.handlers.render_pre.append(renderPre)
 
+
     global addonChanged
     addonChanged = True
 
 def unregister():
     bpy.app.handlers.frame_change_post.remove(frameChangedPost)
+    bpy.app.handlers.depsgraph_update_post.remove(depsgraphUpdatePost)
     bpy.app.handlers.load_post.remove(loadPost)
+    bpy.app.handlers.version_update.remove(versionUpdate)
     bpy.app.handlers.save_pre.remove(savePre)
     bpy.app.timers.unregister(always)